use KeyPad + fixes
added KeyPad lib
speedup LCD4Bit_mod lib

+++ src/LCD4Bit_mod.h
@@ -5,23 +5,23 @@
 
 class LCD4Bit_mod {
 public:
-  LCD4Bit_mod(int num_lines);
-  void commandWrite(int value);
+  LCD4Bit_mod(uint8_t num_lines);
+  void commandWrite(uint8_t value);
   void init();
-  void print(int value);
+  void print(uint8_t value);
   void printIn(const char* msg);
   void clear();
   //non-core---------------
-  void cursorTo(int line_num, int x);
-  void leftScroll(int chars, int delay_time);
+  void cursorTo(uint8_t line_num, uint8_t x);
+  void leftScroll(uint8_t chars, uint8_t delay_time);
   //end of non-core--------
 
   //4bit only, therefore ideally private but may be needed by user
-  void commandWriteNibble(int nibble);
+  void commandWriteNibble(uint8_t nibble);
 private:
   void pulseEnablePin();
-  void pushNibble(int nibble);
-  void pushByte(int value);
+  void pushNibble(uint8_t nibble);
+  void pushByte(uint8_t value);
 };
 
 #endif
